+++ bricks/advanced/src/workbench-layout-v2/DroppableComponentLayoutItem.tsx
@@ -47,6 +47,7 @@ export function DroppableComponentLayoutItemElement(
       {
         ...layout,
         cardWidth: layout.cardWidth || component.position.w,
+        cardTitle: component.title,
       },
       position
     )
@@ -77,14 +78,17 @@ export function DroppableComponentLayoutItemElement(
   const editMaskTop = useMemo(() => {
     return 45 + cardBorderWidth;
   }, [cardBorderWidth]);
+  const noPadding = useMemo(() => {
+    return layout.noPadding;
+  }, [layout]);
 
   const deleteIconPosition = useMemo(() => {
     const top = 15 + cardBorderWidth;
     return {
-      top,
+      top: !noPadding ? top : 0,
       right: top,
     };
-  }, [cardBorderWidth]);
+  }, [cardBorderWidth, noPadding]);
 
   return (
     <div className={styles.itemWrapper}>