+++ bricks/ai-portal/src/cruise-canvas/index.tsx
@@ -21,6 +21,7 @@ import ResizeObserver from "resize-observer-polyfill";
 import { select, type Selection, type TransitionLike } from "d3-selection";
 import { ZoomTransform } from "d3-zoom";
 import { mergeRects } from "@next-shared/diagram";
+import type { ConstructResult } from "@next-shared/jsx-storyboard";
 import { NS, locales } from "./i18n.js";
 import styles from "./styles.module.css";
 import { useZoom } from "./useZoom.js";
@@ -65,9 +66,10 @@ import { Nav } from "./Nav/Nav.js";
 import { ReplayToolbar } from "../shared/ReplayToolbar/ReplayToolbar.js";
 import { ChatBox } from "../shared/ChatBox/ChatBox.js";
 import { FilePreview } from "./FilePreview/FilePreview.js";
-import { NodeFeedback } from "./NodeFeedback/NodeFeedback.js";
+import { NodeFeedback } from "../shared/NodeFeedback/NodeFeedback.js";
 import { TaskContext } from "../shared/TaskContext.js";
 import { NodeLoading } from "./NodeLoading/NodeLoading.js";
+import { JsxEditor } from "../shared/JsxEditor/JsxEditor.js";
 
 initializeI18n(NS, locales);
 
@@ -84,6 +86,8 @@ export interface CruiseCanvasProps {
   supports?: Record<string, boolean>;
   showHiddenJobs?: boolean;
   showFeedback?: boolean;
+  showUiSwitch?: boolean;
+  showJsxEditor?: boolean;
 }
 
 const CruiseCanvasComponent = forwardRef(LegacyCruiseCanvasComponent);
@@ -127,6 +131,15 @@ class CruiseCanvas extends ReactNextElement implements CruiseCanvasProps {
   @property({ type: Boolean })
   accessor showFeedback: boolean | undefined;
 
+  @property({ type: Boolean })
+  accessor showUiSwitch: boolean | undefined;
+
+  @property({ type: Boolean, render: false })
+  accessor hideMermaid: boolean | undefined;
+
+  @property({ type: Boolean })
+  accessor showJsxEditor: boolean | undefined;
+
   @event({ type: "share" })
   accessor #shareEvent!: EventEmitter<void>;
 
@@ -162,6 +175,13 @@ class CruiseCanvas extends ReactNextElement implements CruiseCanvasProps {
     this.#feedbackSubmitEvent.emit(detail);
   };
 
+  @event({ type: "ui.switch" })
+  accessor #switch!: EventEmitter<"chat">;
+
+  #onSwitchToChat = () => {
+    this.#switch.emit("chat");
+  };
+
   #ref = createRef<CruiseCanvasRef>();
 
   @method()
@@ -190,11 +210,14 @@ class CruiseCanvas extends ReactNextElement implements CruiseCanvasProps {
         supports={this.supports}
         showHiddenJobs={this.showHiddenJobs}
         showFeedback={this.showFeedback}
+        showUiSwitch={this.showUiSwitch}
+        showJsxEditor={this.showJsxEditor}
         onShare={this.#onShare}
         onPause={this.#onPause}
         onResume={this.#onResume}
         onCancel={this.#onCancel}
         onSubmitFeedback={this.#onSubmitFeedback}
+        onSwitchToChat={this.#onSwitchToChat}
         ref={this.#ref}
       />
     );
@@ -207,6 +230,7 @@ interface CruiseCanvasComponentProps extends CruiseCanvasProps {
   onResume: () => void;
   onCancel: () => void;
   onSubmitFeedback: (detail: FeedbackDetail) => void;
+  onSwitchToChat: () => void;
 }
 
 interface ScrollToOptions {
@@ -242,11 +266,14 @@ function LegacyCruiseCanvasComponent(
     supports,
     showHiddenJobs,
     showFeedback: propShowFeedback,
+    showUiSwitch,
+    showJsxEditor,
     onShare,
     onPause,
     onResume,
     onCancel,
     onSubmitFeedback,
+    onSwitchToChat,
   }: CruiseCanvasComponentProps,
   ref: React.Ref<CruiseCanvasRef>
 ) {
@@ -669,8 +696,20 @@ function LegacyCruiseCanvasComponent(
   const [activeExpandedViewJobId, setActiveExpandedViewJobId] = useState<
     string | null
   >(null);
-
   const [activeFile, setActiveFile] = React.useState<FileInfo | null>(null);
+  const [activeJsxEditorJob, setActiveJsxEditorJob] = useState<
+    Job | undefined
+  >();
+  const [manuallyUpdatedViews, setManuallyUpdatedViews] = useState<
+    Map<string, ConstructResult> | undefined
+  >();
+  const updateView = useCallback((jobId: string, view: ConstructResult) => {
+    setManuallyUpdatedViews((prev) => {
+      const next = new Map(prev);
+      next.set(jobId, view);
+      return next;
+    });
+  }, []);
 
   const taskContextValue = useMemo(
     () => ({
@@ -680,10 +719,22 @@ function LegacyCruiseCanvasComponent(
       onResume,
       onCancel,
       supports,
+
       activeExpandedViewJobId,
       setActiveExpandedViewJobId,
       activeToolCallJobId,
       setActiveToolCallJobId,
+
+      submittingFeedback,
+      submittedFeedback,
+      onSubmitFeedback: handleSubmitFeedback,
+      setShowFeedback,
+
+      showJsxEditor,
+      activeJsxEditorJob,
+      setActiveJsxEditorJob,
+      manuallyUpdatedViews,
+      updateView,
     }),
     [
       humanInput,
@@ -692,30 +743,30 @@ function LegacyCruiseCanvasComponent(
       onResume,
       onShare,
       supports,
+
       activeExpandedViewJobId,
       activeToolCallJobId,
+
+      submittingFeedback,
+      submittedFeedback,
+      handleSubmitFeedback,
+
+      showJsxEditor,
+      activeJsxEditorJob,
+      manuallyUpdatedViews,
+      updateView,
     ]
   );
 
   const canvasContextValue = useMemo(
     () => ({
       onNodeResize,
-      onSubmitFeedback: handleSubmitFeedback,
       setActiveNodeId,
       hoverOnScrollableContent,
       setHoverOnScrollableContent,
       setActiveFile,
-      setShowFeedback,
-      submittingFeedback,
-      submittedFeedback,
     }),
-    [
-      hoverOnScrollableContent,
-      onNodeResize,
-      handleSubmitFeedback,
-      submittingFeedback,
-      submittedFeedback,
-    ]
+    [hoverOnScrollableContent, onNodeResize]
   );
 
   const activeToolCallJob = useMemo(() => {
@@ -901,8 +952,10 @@ function LegacyCruiseCanvasComponent(
           />
           <ZoomBar
             scale={transform.k}
+            showUiSwitch={showUiSwitch}
             onScaleChange={handleScaleChange}
             onReCenter={handleReCenter}
+            onSwitchToChat={onSwitchToChat}
           />
           {replay ? (
             <div className={styles["footer-container"]}>
@@ -924,6 +977,7 @@ function LegacyCruiseCanvasComponent(
         {activeToolCallJob && <ToolCallDetail job={activeToolCallJob} />}
         {activeExpandedViewJobId && <ExpandedView views={views!} />}
         {activeFile && <FilePreview file={activeFile} />}
+        {showJsxEditor && activeJsxEditorJob && <JsxEditor />}
       </CanvasContext.Provider>
     </TaskContext.Provider>
   );