+++ bricks/ai-portal/src/elevo-sidebar/index.tsx
@@ -3,7 +3,9 @@ import React, {
   forwardRef,
   useCallback,
   useEffect,
+  useImperativeHandle,
   useMemo,
+  useRef,
   useState,
 } from "react";
 import { createDecorators, type EventEmitter } from "@next-core/element";
@@ -59,6 +61,10 @@ export interface SidebarLink {
 
 const ElevoSidebarComponent = forwardRef(LegacyElevoSidebarComponent);
 
+interface ElevoSidebarRef extends ChatHistoryRef {
+  close: () => void;
+}
+
 /**
  * 构件 `ai-portal.elevo-sidebar`
  */
@@ -105,17 +111,22 @@ class ElevoSidebar extends ReactNextElement implements ElevoSidebarProps {
     this.#actionClick.emit(detail);
   };
 
-  #ref = createRef<ChatHistoryRef>();
+  #ref = createRef<ElevoSidebarRef>();
 
   /**
    * @param delay Delay in milliseconds before pulling the latest chat history.
    */
   @method()
   pullHistory(delay: number) {
-    // Wait 3 seconds to let the task title to be summarized.
+    // Wait several seconds to let the task title to be summarized.
     setTimeout(() => {
       this.#ref.current?.pull();
     }, delay);
+  }
+
+  @method()
+  close() {
+    this.#ref.current?.close();
   }
 
   render() {
@@ -155,7 +166,7 @@ function LegacyElevoSidebarComponent(
     onLogout,
     onActionClick,
   }: ElevoSidebarComponentProps,
-  ref: React.Ref<ChatHistoryRef>
+  ref: React.Ref<ElevoSidebarRef>
 ) {
   const [collapsed, setCollapsed] = useState(behavior === "drawer");
   const handleCollapse = useCallback(() => {
@@ -204,11 +215,41 @@ function LegacyElevoSidebarComponent(
     []
   );
 
+  const historyRef = useRef<ChatHistoryRef>(null);
+
+  useImperativeHandle(
+    ref,
+    () => ({
+      close: () => {
+        setCollapsed(true);
+      },
+      pull: () => {
+        historyRef.current?.pull();
+      },
+    }),
+    []
+  );
+
   return (
     <div className={classNames("container", { collapsed })}>
       {behavior === "drawer" && !collapsed && (
         <div className="mask" onClick={handleClickMask} />
       )}
+      <div className="alternative">
+        <WrappedIconButton
+          icon={SIDEBAR_ICON}
+          variant="light"
+          onClick={handleExpand}
+        />
+        <WrappedLink className="new-chat" url={newChatUrl}>
+          <WrappedIcon
+            className="new-chat-icon"
+            lib="easyops"
+            icon="new-chat"
+          />
+          {t(K.NEW_CHAT)}
+        </WrappedLink>
+      </div>
       <div className="sidebar">
         <div className="logo-bar">
           <WrappedLink url={logoUrl} className="logo-link">
@@ -217,7 +258,7 @@ function LegacyElevoSidebarComponent(
               alt="Elevo"
               src={ElevoLogo}
               width={95}
-              height={26}
+              height={28}
             />
           </WrappedLink>
           <WrappedIconButton
@@ -245,7 +286,7 @@ function LegacyElevoSidebarComponent(
           </div>
         ) : null}
         <ChatHistory
-          ref={ref}
+          ref={historyRef}
           activeId={historyActiveId}
           urlTemplate={historyUrlTemplate}
           actions={historyActions}
@@ -278,21 +319,6 @@ function LegacyElevoSidebarComponent(
             </button>
           </WrappedDropdownActions>
         </div>
-      </div>
-      <div className="alternative">
-        <WrappedIconButton
-          icon={SIDEBAR_ICON}
-          variant="light"
-          onClick={handleExpand}
-        />
-        <WrappedLink className="new-chat" url={newChatUrl}>
-          <WrappedIcon
-            className="new-chat-icon"
-            lib="easyops"
-            icon="new-chat"
-          />
-          {t(K.NEW_CHAT)}
-        </WrappedLink>
       </div>
     </div>
   );