+++ bricks/ai-portal/src/shared/CreatedView/CreatedView.tsx
@@ -34,12 +34,17 @@ export function CreatedView({
   onSizeChange,
 }: CreatedViewProps): JSX.Element {
   const rootId = useMemo(() => uniqueId(), []);
-  const { setActiveExpandedViewJobId } = useContext(TaskContext);
+  const {
+    showJsxEditor,
+    setActiveExpandedViewJobId,
+    setActiveJsxEditorJob,
+    manuallyUpdatedViews,
+  } = useContext(TaskContext);
   const ref = useRef<HTMLDivElement>(null);
   const rootRef = useRef<Awaited<
     ReturnType<typeof unstable_createRoot>
   > | null>(null);
-  const view = job.generatedView!;
+  const view = manuallyUpdatedViews?.get(job.id) ?? job.generatedView!;
 
   useEffect(() => {
     const container = ref.current;
@@ -92,10 +97,13 @@ export function CreatedView({
     if (isJsxView(view)) {
       // TODO: handle nested components
       for (const component of view.components ?? []) {
-        if (component.name === "table") {
+        if (component.name === "Table" || component.name === "eo-table") {
           return true;
         }
-        if (component.name === "dashboard") {
+        if (
+          component.name === "Dashboard" ||
+          component.name === "eo-dashboard"
+        ) {
           const widgets = component?.properties?.widgets;
           if (
             Array.isArray(widgets) &&
@@ -142,13 +150,25 @@ export function CreatedView({
         >
           {view.title}
         </div>
-        <button
-          className={styles.expand}
-          title={t(K.FULLSCREEN)}
-          onClick={handleExpandClick}
-        >
-          <WrappedIcon lib="easyops" icon="expand" />
-        </button>
+        <div className={styles.buttons}>
+          {showJsxEditor && isJsxView(view) && (
+            <button
+              className={styles.button}
+              onClick={() => {
+                setActiveJsxEditorJob?.(job);
+              }}
+            >
+              <WrappedIcon lib="antd" icon="bug" />
+            </button>
+          )}
+          <button
+            className={styles.button}
+            title={t(K.FULLSCREEN)}
+            onClick={handleExpandClick}
+          >
+            <WrappedIcon lib="easyops" icon="expand" />
+          </button>
+        </div>
       </div>
       <div data-root-id={rootId} ref={ref} />
     </div>