+++ bricks/vb-experiment/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.49](https://github.com/easyops-cn/next-advanced-bricks/compare/@next-bricks/vb-experiment@1.0.48...@next-bricks/vb-experiment@1.0.49) (2025-08-20)
+
+
+### Bug Fixes
+
+* support converting tsx storyboard ([f4809f2](https://github.com/easyops-cn/next-advanced-bricks/commit/f4809f26239f782465aa9c8678e50094c68bc4fd))
+
+
+
+
+
 ## [1.0.48](https://github.com/easyops-cn/next-advanced-bricks/compare/@next-bricks/vb-experiment@1.0.47...@next-bricks/vb-experiment@1.0.48) (2025-07-10)
 
 **Note:** Version bump only for package @next-bricks/vb-experiment