+++ bricks/visual-builder/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.39.3](https://github.com/easyops-cn/next-advanced-bricks/compare/@next-bricks/visual-builder@1.39.2...@next-bricks/visual-builder@1.39.3) (2025-08-14)
+
+**Note:** Version bump only for package @next-bricks/visual-builder
+
+
+
+
+
+## [1.39.2](https://github.com/easyops-cn/next-advanced-bricks/compare/@next-bricks/visual-builder@1.39.1...@next-bricks/visual-builder@1.39.2) (2025-08-13)
+
+**Note:** Version bump only for package @next-bricks/visual-builder
+
+
+
+
+
 ## [1.39.1](https://github.com/easyops-cn/next-advanced-bricks/compare/@next-bricks/visual-builder@1.39.0...@next-bricks/visual-builder@1.39.1) (2025-07-29)
 
 **Note:** Version bump only for package @next-bricks/visual-builder